News_fetcher/requirements.txt
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0
google-genai>=1.30,<2.0
